fix(hover): show hover info only at the begin of each line
The hover infos are only shown within the first 21 chars of each line.
Rationale: as each line contains a hover info and the info is not column specific
its a bit annoying on scrolling in longer texts.
If you dislike this change please file an issue and I'll make this a
configurable option.