fix(restquery): prefer the docs that are visible
If the get/docs/<id> id is not the document path but a number (e.g. 0)
the id'th one from the tree view - so the visible dlt docs - is used and not
from the known/still in mem documents.
Fallback to prev method if e.g. no document is visble.
Added a deprecated warning for the case where more than 1 document is know.
This fallback should be removed in the future.