fix(lifecycledetection): limit the number of logs
Limit the number of two log output as they can occur really often if a weird
problem occurs:
the traces suddenly contain one completely different msg (from e.g. two days later)
and then the regular messages. In those cases the lifecycle detection keeps them as part
of the current lifecycle (that was started with the strange message) and
logs a lot of warnings on console.
Limited at least that as a hotfix.