feat(restquery): allow queries with more than 1000 msgs as result
A filter can now contain the "maxNrMsgs" key.
It can either be a positive number or 0 for all.
If a pos. number is provided the maximum of 1000 and all provided numbers is used.
If 0 is used the max is set to unlimited.
This should be used with caution and only in cases where e.g. the result of
all msgs is needed and processed afterwards e.g. with a javascript function.