fix(config): update config setting where it currently exists
To fix #13 where the filters are determined by the workspace config and not the
global (aka user) config we do update the settings
where its currently found.
This might "copy" more than wanted to e.g. workspace as getConfiguration merges the
settings but we will then persist the full object in the highest order
target/location.

Issue: #13