feat: thin out data on timeline charts
If a single line in a timeline chart contains more than 500 items
the data gets thinned out to around half of the width in pixel.
On zoom, pan resize the thinning is redone so that e.g. on a high zoom
level all data is shown.
This speeds up reports with e.g. >10k items per line significantly.