feat(export): use less memory, filter on pass 1
Even though the export was already memory optimized it still kept for
every log in the input files a small object.
This worked up to around 40mio logs.
For bigger files it ran out of mem.
Optimized this now by filtering msgs with some heuristics already at
the first pass. This is a bit tricky as the lifecycle calculation usually
needs a lot of logs to be available to be stable.
Thus doing it a bit carefully e.g. not on the last 1mio logs.
For time filters a similar problem exists. Thus using a window of 0/120s
for messages before/after the set filter criteria.
The new approach still keeps one big array with either the min log data or
undefined. So it should work up to at least around 10 times the amount of
input data if filter criteria are used that output only around 20mio logs.