feat(someip): add report for methods, change fire&forget symbol
Add report for someip methods from tree view.
Change someip requests wo response (fire&forget) call symbol to '\' instead of '>'.
'>' is used for requests that expect a response.
'<' is used for the response.

This might be a breaking change for any filters that rely on '> ...'!.
You can change them to "^[><\\] " to get all someip methods wo error messages.