feat(comment): first minimal version of comments
This allows to add comments (using the small '+' icon at the start of every log line)
to a single log or a selected block of logs.
The comments can then be exported to the clipboard in Markup code, i.e.
the log lines will be embedded in a {noformat} block.
Currently comments are persisted only in the globalState of the extension
so only on the local machine.
As storage key the URI (so indirectly the path/name of the files) are used.
They comments are purged/deleted after >30days.