feat(search): cmd/ctrl+a/c copies all search results to clipboard
As the search window is using a windowed list copying "all" content
from there does not really copy all but only the shown/windowed ones.

Removed the default behaviour (context menu and ctrl/cmd+a, ctrl/cmd+c)
and replaced by an own implementation that really copies the full result to clipboard.
If not all messages are selected (via ctrl/cmd+a) then the default behaviour is kept.