feat(search): set/deactive filters directly from search
Added two new activity buttons to the search
panel: '+' and '-'.
'+' will add the current search as a filter
  (or activate any existing ones)
'-' will deactivate the current search as a filter (if a filter is currently set).

An added filter is not persisted to configuration but can be edited from
the tree view/filter and will then be persisted.