fix: hotfix to prevent .git/config files opened as document
Seems some extensions query the filesystem provider
with readFile of .git/config files.
This is causing issues 'config' file shown in tree view
and search/restquery not working
as config is then the active file.

Hotfix is to avoid opening uri with files inside archives
that do not locally exist.
So opening any archive file that contain '.git/config' is currently not possible.
Proper fix to handle the situation where a non existing relative path
inside an archive is used for stat/readFile is needed.

Provided a clean solution for the Buffer.from warning as well
via TextEncoder.